Move HtmlTextUtils to services directory
